[WIP] Adding test framework
Use 'npm test' command to run the tests (uses jest framework).

Something I'm still trying to figure out is how to mock the "tiled" global, and persist changes to it (e.g., to set the activeAsset in the "arrange" section of the tests, and have those changes persist when calling methods on BulkAnimationEditor).

It's also a bit of a challenge to import the BulkAnimationEditor class in tests, since I am not able to export it without breaking when the extension is loaded within the Tiled editor. Normally, one would use the "export" keyword, or use module.exports - however, this results in errors when running the extension (and I can't check to see if module is defined, since all of window, global, globalThis, and self are undefined and throw reference errors when it is loaded in Tiled). I am using the rewire package to be able to get at the non-exported bulkAnimationEditor variable, which seems to work fine - however, it seems this results in multiple instances of the "tiled" global, which is a bit of a problem since I need to be able to mutate it from within tests.